move templates to separate file
